#59 Kompatibel mit Postgres machen

+++ migrations/postgres/20201103010530_create_roles.sql
@@ -1,14 +1,14 @@
 -- Your SQL goes here
 CREATE TABLE roles
 (
-    id          SERIAL       NOT NULL PRIMARY KEY,
+    id          BIGSERIAL    NOT NULL PRIMARY KEY,
     name        VARCHAR(128) NOT NULL UNIQUE,
     permissions TEXT         NOT NULL DEFAULT ''
 );
 
 CREATE TABLE users_roles
 (
-    id      SERIAL  NOT NULL PRIMARY KEY,
-    user_id INTEGER NOT NULL REFERENCES users (id) ON UPDATE CASCADE ON DELETE CASCADE,
-    role_id INTEGER NOT NULL REFERENCES roles (id) ON UPDATE CASCADE
+    id      BIGSERIAL NOT NULL PRIMARY KEY,
+    user_id BIGINT    NOT NULL REFERENCES users (id) ON UPDATE CASCADE ON DELETE CASCADE,
+    role_id BIGINT    NOT NULL REFERENCES roles (id) ON UPDATE CASCADE
 );
